perf: add PathSum table (#11)
Summing over 150M rows in `path_qtys` is slow. One solution is a "materialized view", but updating such a view requires updating *all* rows in the view even when only a few need to be added. Therefore a `path_sums` table is introduced, which contains buy+sell quantity for all paths for a calculation. A row is inserted into this table for a calculation when the paths and quantities for this calculation are inserted.